# docs/swagger.yaml
definitions:
...
  entity.VietQR:
    properties:
      id:
        type: string
      status:
        type: string
    type: object
...
  request.UpdateVietQRStatus:
    properties:
      status:
        type: string
    required:
    - status
    type: object
  response.Error:
    properties:
      error:
        example: message
        type: string
    type: object
...
  response.Success:
    properties:
      message:
        example: operation completed successfully
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Using a translation service and user management as examples
  title: Go Dev Kit Template API
  version: "1.0"
paths:
...
  /v1/vietqr/gen:
    post:
      consumes:
      - application/json
      description: Generate a new VietQR code
      operationId: generate-qr
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.VietQR'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Error'
      security:
      - BearerAuth: []
      summary: Generate QR Code
      tags:
      - vietqr
  /v1/vietqr/inquiry/{id}:
    get:
      consumes:
      - application/json
      description: Get the status of a VietQR code
      operationId: inquiry-qr
      parameters:
      - description: QR ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.VietQR'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Error'
      security:
      - BearerAuth: []
      summary: Inquiry QR Status
      tags:
      - vietqr
  /v1/vietqr/update/{id}:
    put:
      consumes:
      - application/json
      description: Update the status of a VietQR code
      operationId: update-qr-status
      parameters:
      - description: QR ID
        in: path
        name: id
        required: true
        type: string
      - description: Update status
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.UpdateVietQRStatus'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Error'
      security:
      - BearerAuth: []
      summary: Update QR Status
      tags:
      - vietqr
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"